Create HTML for todos
working on creating a working version of the todo list. It is not quiet functional yet it is not displaying, but I am suspecting it to be a routing issue.  Also made status into its own type since I remember that with the server we had to select between complete and incomplete